+++ R/ooxml.R
@@ -451,6 +451,13 @@ rpr_pml <- function(x) {
     }
   }
 
+  has_strike <- !is.null(x$strike) && !is.na(x$strike) && x$strike
+  if (has_strike) {
+    out <- paste0(out, " strike=\"sngStrike\"")
+  } else {
+    out <- paste0(out, " strike=\"noStrike\"")
+  }
+
   if (x$vertical.align == "superscript") {
     out <- paste0(out, " baseline=\"40000\"")
   } else if (x$vertical.align == "subscript") {
@@ -518,6 +525,13 @@ rpr_wml <- function(x) {
     } else {
       out <- paste0(out, "<w:u w:val=\"none\"/>")
     }
+  }
+
+  has_strike <- !is.null(x$strike) && !is.na(x$strike) && x$strike
+  if (has_strike) {
+    out <- paste0(out, "<w:strike w:val=\"true\"/>")
+  } else {
+    out <- paste0(out, "<w:strike w:val=\"false\"/>")
   }
 
   if (x$vertical.align == "superscript") {
@@ -596,12 +610,14 @@ rpr_css <- function(x) {
     }
   }
 
-  if (!is.na(x$bold)) {
-    if (x$underlined) {
-      out <- paste0(out, "text-decoration:underline;")
-    } else {
-      out <- paste0(out, "text-decoration:none;")
-    }
+  has_underline <- !is.na(x$underlined) && x$underlined
+  if (has_underline) {
+    out <- paste0(out, "text-decoration:underline;")
+  }
+
+  has_strike <- !is.null(x$strike) && !is.na(x$strike) && x$strike
+  if (has_strike) {
+    out <- paste0(out, "text-decoration:line-through;")
   }
 
   if (!is.na(x$shading.color)) {